+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog Roundcube Webmail
 
+## Unreleased
+
+- Support displaying RTF content (including encapsulated HTML) from a TNEF attachment
+
 ## Release 1.5-rc
 
 - Upgrade to TinyMCE 5.8.2